feat: separate seaware api client
